update application properties, build configuration, and refactor SQL tool integration for version 1.0.0
