+++ chapter5/1-natural-sql-llm-functions/build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'apps'
-version = '0.0.1-SNAPSHOT'
+version = '0.0.1'
 
 java {
     sourceCompatibility = '17'
@@ -23,7 +23,7 @@ repositories {
 }
 
 ext {
-    set('springAiVersion', "1.0.0-SNAPSHOT")
+    set('springAiVersion', "1.0.0")
     set('springCloudGcpVersion', "5.4.3")
     set('springCloudVersion', "2023.0.1")
 }
@@ -34,7 +34,7 @@ dependencies {
 
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
     implementation 'com.google.cloud:spring-cloud-gcp-starter'
-    implementation 'org.springframework.ai:spring-ai-vertex-ai-gemini-spring-boot-starter'
+    implementation 'org.springframework.ai:spring-ai-starter-model-vertex-ai-gemini'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
