+++ chapter5/1-natural-sql-llm-functions/src/main/java/apps/naturalsqlai/NaturalSqlAiApplication.java
@@ -1,13 +1,8 @@
 package apps.naturalsqlai;
 
-import apps.naturalsqlai.products.sql.LLMSqlQueryExecutor;
-import apps.naturalsqlai.products.sql.QueryRequest;
-import apps.naturalsqlai.products.sql.QueryResponse;
-import org.springframework.ai.model.function.FunctionCallbackWrapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.ai.model.function.FunctionCallbackContext; 
+
 
 @SpringBootApplication
 public class NaturalSqlAiApplication {
@@ -15,16 +10,4 @@ public class NaturalSqlAiApplication {
     public static void main(String[] args) {
         SpringApplication.run(NaturalSqlAiApplication.class, args);
     }
-
-
-    @Bean
-    public FunctionCallbackWrapper<QueryRequest, QueryResponse> llmSqlQueryExecutorFunction(LLMSqlQueryExecutor llmSqlQueryExecutor) {
-        return FunctionCallbackWrapper.builder(llmSqlQueryExecutor)
-                .withName("execute_sql_queries")
-                .withSchemaType(FunctionCallbackContext.SchemaType.OPEN_API_SCHEMA)
-                .withDescription("Execute SQL queries")
-                .build();
-    }
-
-
 }
