+++ chapter5/1-natural-sql-llm-functions/src/main/java/apps/naturalsqlai/products/ProductController.java
@@ -2,19 +2,24 @@ package apps.naturalsqlai.products;
 
 import apps.naturalsqlai.ChatBotRequest;
 import apps.naturalsqlai.ChatBotResponse;
-import apps.naturalsqlai.products.sql.QueryRequest;
-import apps.naturalsqlai.products.sql.QueryResponse;
+import apps.naturalsqlai.products.tools.SQLTools;
+import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.ai.chat.client.ChatClient;
+import org.springframework.ai.chat.messages.AssistantMessage;
 import org.springframework.ai.chat.messages.Message;
 import org.springframework.ai.chat.messages.SystemMessage;
 import org.springframework.ai.chat.messages.UserMessage;
 import org.springframework.ai.chat.model.ChatResponse;
 import org.springframework.ai.chat.prompt.ChatOptions;
 import org.springframework.ai.chat.prompt.Prompt;
-import org.springframework.ai.model.function.FunctionCallbackWrapper;
+import org.springframework.ai.model.tool.ToolCallingChatOptions;
+import org.springframework.ai.support.ToolCallbacks;
+import org.springframework.ai.tool.ToolCallback;
 import org.springframework.ai.vertexai.gemini.VertexAiGeminiChatModel;
 import org.springframework.ai.vertexai.gemini.VertexAiGeminiChatOptions;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,7 +29,6 @@ import java.io.BufferedReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.stream.Collectors;
 
 @RestController
@@ -34,7 +38,18 @@ import java.util.stream.Collectors;
 public class ProductController {
 
     private final VertexAiGeminiChatModel vertexAiGeminiChatModel;
-    private final FunctionCallbackWrapper<QueryRequest, QueryResponse> llmSqlQueryExecutorFunction;
+
+    private final JdbcTemplate jdbcTemplate;
+
+    private ChatClient chatClient;
+
+    @PostConstruct
+    void setup() {
+        chatClient = ChatClient.builder(vertexAiGeminiChatModel)
+                .defaultSystem("You are a smart helpful assistant that can answer user questions, generate SQL query/queries for that user question, execute those SQL query/queries and respond in human understandable language.")
+                .defaultTools(new SQLTools(jdbcTemplate))
+                .build();
+    }
 
 
     public String readFromClasspath(String filename) {
@@ -55,26 +70,24 @@ public class ProductController {
         String question = chatBotRequest.question();
         ;
 
-        String context = readFromClasspath("generate-sql-function-prompt.txt");
+        String context = readFromClasspath("generate-sql-tool-calling-prompt.txt");
+        String assistantMessage = readFromClasspath("assistant-prompt-sql-tool-calling.txt");
 
 
         var messages = new ArrayList<Message>();
-        messages.add(new SystemMessage(context));
+        messages.add(new UserMessage(context));
+        messages.add(new AssistantMessage(assistantMessage));
         messages.add(new UserMessage(question));
 
 
-        ChatOptions options = VertexAiGeminiChatOptions.builder()
-                .withFunctionCallbacks(List.of(llmSqlQueryExecutorFunction))
-                .build();
+        Prompt prompt = new Prompt(messages);
 
-        Prompt prompt = new Prompt(messages, options);
         // call the chat client
-        ChatResponse chatResponse = vertexAiGeminiChatModel.call(prompt);
+        String answer = chatClient.prompt(prompt)
+                                        .call()
+                                        .content();
 
-        log.info("Response: {}", chatResponse);
-        // get the answer
-        String answer = chatResponse.getResult().getOutput().getContent();
-
+        log.info("Response: {}", answer);
 
         return new ChatBotResponse(question, answer);
