--- chapter5/1-natural-sql-llm-functions/src/main/java/apps/naturalsqlai/products/sql/QueryRequest.java
@@ -1,12 +1,0 @@
-package apps.naturalsqlai.products.sql;
-
-import com.fasterxml.jackson.annotation.JsonClassDescription;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyDescription;
-
-import java.util.List;
-
-@JsonClassDescription("Execute SQL queries request")
-public record QueryRequest(@JsonProperty(required = true, value = "sqls")
-                           @JsonPropertyDescription("The query to be executed as list of sqls ")  List<String> sqls) {
-}
