+++ chapter5/1-natural-sql-llm-functions/src/main/resources/application.properties
@@ -12,7 +12,7 @@ spring.ai.vertex.ai.gemini.location=us-central1
 ## Identify the models from https://console.cloud.google.com/vertex-ai/model-garden
 #spring.ai.vertex.ai.gemini.chat.options.model=gemini-1.5-flash
 
-spring.ai.vertex.ai.gemini.chat.options.model=gemini-1.0-pro
+spring.ai.vertex.ai.gemini.chat.options.model=gemini-2.5-pro-preview-05-06
 #spring.ai.vertex.ai.gemini.chat.options.model=gemini-1.5-pro
 
 ### Local related properties
